# Optimize comparison overhead from `_wrap_type_error_to_not_implemented`