One consequence of https://github.com/pytorch/pytorch/pull/37091 is that comparison operators have to check for `__torch_function__`. However the import is called each time as part of the hot path, and while Python is lazy about re-import it is not free. This PR simply moves the import up one level to decorator construction time, which saves about 1.5 us:

```
from torch.utils.benchmark import Timer

timer = Timer(
    "x == y",
    "x = torch.ones((1,)); y = torch.ones((1,))",
)

print(timer.blocked_autorange(min_run_time=1), "\n\n")
print(timer.collect_callgrind())
```

Before:
```
<torch.utils.benchmark.utils.common.Measurement object at 0x7fb9adec3d30>
x == y
setup: x = torch.ones((1,)); y = torch.ones((1,))
  Median: 7.00 us
  IQR:    0.27 us (6.89 to 7.16)
  143 measurements, 1000 runs per measurement, 1 thread


<torch.utils.benchmark.utils.valgrind_wrapper.timer_interface.CallgrindStats object at 0x7fb9aded0400>
x == y
setup: x = torch.ones((1,)); y = torch.ones((1,))
                           All          Noisy symbols removed
    Instructions:      2499971                    2373429
    Baseline:             4401                       3981
100 runs per measurement, 1 thread
```

After:
```
<torch.utils.benchmark.utils.common.Measurement object at 0x7f966fd02898>
x == y
setup: x = torch.ones((1,)); y = torch.ones((1,))
  Median: 5.43 us
  IQR:    0.25 us (5.34 to 5.59)
  177 measurements, 1000 runs per measurement, 1 thread


<torch.utils.benchmark.utils.valgrind_wrapper.timer_interface.CallgrindStats object at 0x7f966fd02ef0>
x == y
setup: x = torch.ones((1,)); y = torch.ones((1,))
                           All          Noisy symbols removed
    Instructions:      1759233                    1705173
    Baseline:             4335                       3954
100 runs per measurement, 1 thread
```

This drops instruction count by ~40% and overhead wall time by ~30%. Otherwise it should be a no-op. The check itself adds a further ~1us over the old state, but that is much harder to claw back.

Test plan: Existing unit tests should be sufficient.