@robieta 2020-11-16
Closing in favor of https://github.com/pytorch/pytorch/pull/48018, which eschews Python altogether.